+++ diive/__init__.py
@@ -1,19 +1,17 @@
 from diive.core.dfun.frames import transform_yearmonth_matrix_to_longform as transform_yearmonth_matrix_to_longform
-
-# Plotting
+# Heatmap
 from diive.core.plotting.heatmap_datetime import HeatmapDateTime as heatmapdatetime
 from diive.core.plotting.heatmap_datetime import HeatmapYearMonth as heatmapyearmonth
 from diive.core.plotting.heatmap_datetime import HeatmapYearMonthRanks as heatmapyearmonth_ranks
+# Ridgeline
 from diive.core.plotting.ridgeline import RidgeLinePlot as ridgeline
-
 from diive.core.times.resampling import resample_to_monthly_agg_matrix as resample_to_monthly_agg_matrix
 from diive.pkgs.createvar.conversions import et_from_le as et_from_le
 
-
+# Plotting
 
 # import diive.core.plotting as plot
 # plot.ridgeline()
-
 # def ridgeplot(*args, **kwargs):
 #     """A convenience function to create and display RidgePlotTS."""
 #     return RidgePlotTS(*args, **kwargs)