+++ diive/core/plotting/cumulative.py
@@ -168,6 +168,11 @@ class Cumulative:
         if self.start_year | self.end_year:
             self.df = keep_years(data=self.df, start_year=self.start_year, end_year=self.end_year)
 
+        self.show_grid = True
+        self.show_legend = True
+        self.ylabel = None
+        self.show_title = True
+
         self.cumulative = self.df.cumsum()
 
         # Create axis
@@ -175,20 +180,26 @@ class Cumulative:
         self.ax.xaxis.axis_date()
 
     def _apply_format(self):
-        title = f"Cumulatives ({self.cumulative.index.min()}-{self.cumulative.index.max()})"
-        self.fig.suptitle(title, fontsize=theme.FIGHEADER_FONTSIZE)
+        if self.show_title:
+            title = f"Cumulatives ({self.cumulative.index.min()}-{self.cumulative.index.max()})"
+            self.fig.suptitle(title, fontsize=theme.FIGHEADER_FONTSIZE)
         ymin = self.cumulative.min().min()
         ymax = self.cumulative.max().max()
+        ymax = ymax * 1.05 if ymax > 0 else ymax * 0.95
+        ymin = ymin * 0.95 if ymin > 0 else ymin * 1.05
         self.ax.set_ylim(ymin, ymax)
         pf.add_zeroline_y(ax=self.ax, data=self.cumulative)
+        ax_ylabel_txt = self.ylabel if self.ylabel else "Cumulative"
         pf.default_format(ax=self.ax,
                           ax_xlabel_txt="Date",
-                          ax_ylabel_txt="Cumulative",
-                          txt_ylabel_units=self.units)
+                          ax_ylabel_txt=ax_ylabel_txt,
+                          txt_ylabel_units=self.units,
+                          showgrid=self.show_grid)
         n_legend_cols = pf.n_legend_cols(n_legend_entries=len(self.cumulative.columns))
-        pf.default_legend(ax=self.ax,
-                          labelspacing=0.2,
-                          ncol=n_legend_cols)
+        if self.show_legend:
+            pf.default_legend(ax=self.ax,
+                              labelspacing=0.2,
+                              ncol=n_legend_cols)
         pf.nice_date_ticks(ax=self.ax, minticks=3, maxticks=20, which='x', locator='auto')
 
         # self.fig.tight_layout()
@@ -197,7 +208,18 @@ class Cumulative:
         """Return axis"""
         return self.ax
 
-    def plot(self, showplot: bool = True, digits_after_comma: int = 0):
+    def plot(self, showplot: bool = True,
+             digits_after_comma: int = 0,
+             show_grid: bool = True,
+             show_legend: bool = True,
+             ylabel: str = None,
+             show_title: bool = True
+             ):
+        self.show_grid = show_grid
+        self.show_legend = show_legend
+        self.ylabel = ylabel
+        self.show_title = show_title
+
         color_list = theme.colorwheel_36()  # get some colors
 
         # Plot yearly cumulatives
@@ -229,7 +251,7 @@ class Cumulative:
             self.fig.show()
 
 
-def example_cum_overall():
+def _example_cum_overall():
     # Test data
     from diive.configs.exampledata import load_exampledata_parquet
     df_orig = load_exampledata_parquet()
@@ -243,10 +265,18 @@ def example_cum_overall():
         end_year=2019).plot()
 
 
-def example_cum_year():
-    # Test data
-    from diive.configs.exampledata import load_exampledata_parquet
-    df_orig = load_exampledata_parquet()
+def _example_cum_year():
+    # # Test data
+    # from diive.configs.exampledata import load_exampledata_parquet
+    # df_orig = load_exampledata_parquet()
+
+    from pathlib import Path
+    from diive.core.io.filereader import ReadFileType
+    filepath = r"F:\Sync\luhk_work\40 - DATA\DATASETS\2025_FORESTS\1-downloads\ICOSETC_CH-Dav_ARCHIVE_L2\ICOSETC_CH-Dav_FLUXNET_HH_L2.csv"
+    loaddatafile = ReadFileType(filetype='FLUXNET-FULLSET-HH-CSV-30MIN',
+                                filepath=filepath,
+                                data_nrows=None)
+    df_orig, metadata_df = loaddatafile.get_filedata()
 
     df = df_orig.copy()
 
@@ -271,7 +301,7 @@ def example_cum_year():
     # series.index = pd.to_datetime(series.index)
     # series = series.groupby(series.index.year).mean()  # yearly mean
 
-    series = df['NEE_CUT_REF_f'].copy()
+    series = df['GPP_NT_CUT_50'].copy()
     # series = df['NEE_CUT_REF_f'].copy()
     series = series.multiply(0.02161926)  # umol CO2 m-2 s-1 --> g C m-2 30min-1
     series_units = r'($\mathrm{gC\ m^{-2}}$)'
@@ -285,8 +315,8 @@ def example_cum_year():
         series_units=series_units,
         yearly_end_date=None,
         # yearly_end_date='08-11',
-        start_year=2005,
-        end_year=2020,
+        start_year=2020,
+        end_year=2024,
         show_reference=True,
         excl_years_from_reference=None,
         # excl_years_from_reference=[2022],
@@ -295,5 +325,5 @@ def example_cum_year():
 
 
 if __name__ == '__main__':
-    example_cum_overall()
-    # example_cum_year()
+    # _example_cum_overall()
+    _example_cum_year()